Made major changes to structure and model

+++ Assignment 03/src/Log.hpp
@@ -1,40 +1,35 @@
-//
-//  Log.hpp
-//  Assignment 03
-//
-//
+	//
+	//  Log.hpp
+	//  Assignment 03
+	//
+	//
 
 #ifndef Log_hpp
 #define Log_hpp
 
-#include "Entry.hpp"
 #include <stdio.h>
 #include <iostream>
-#include <fstream>
-#include <string>
-#include <vector>
 #include <string>
 
-class Log{
+class Log {
 private:
 	std::string eventName;
 	float value;
-	float weight;
+	int weight;
 	
 public:
 	Log();
-	Log(std::string eventName, float value, float weight);
-	
+	Log(std::string eventName, float value, int weight);
 	Log(std::string eventName, int value, int weight);
 	~Log();
 	
 	std::string getEventName();
 	float getValue();
-	float getWeight();
+	int getWeight();
 	
 	void setEventName(std::string);
 	void setValue(float);
-	void setWeight(float);
+	void setWeight(int);
 };
 
 #endif /* Log_hpp */
